--- README.md
@@ -1,2 +1,0 @@
-# Robots
-The project to learn OO design concepts and MDI application development in Java
