+++ src/main/java/org/example/gui/LogWindow.java
