+++ src/main/java/org/example/gui/MainApplicationFrame.java
@@ -1,0 +1,96 @@
+package gui;
+
+import java.awt.Dimension;
+import java.awt.Toolkit;
+
+import javax.swing.JDesktopPane;
+import javax.swing.JFrame;
+import javax.swing.JInternalFrame;
+
+
+import log.Logger;
+
+
+
+/**
+ * Что требуется сделать:
+ * 1. Метод создания меню перегружен функционалом и трудно читается.
+ * Следует разделить его на серию более простых методов (или вообще выделить отдельный класс).
+ *
+ */
+
+public class MainApplicationFrame extends JFrame  //
+{
+    private final JDesktopPane desktopPane = new JDesktopPane();
+
+    public MainApplicationFrame() {
+        //Make the big window be indented 50 pixels from each edge
+        //of the screen.
+        int inset = 50;
+        Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
+        setBounds(inset, inset,
+            screenSize.width  - inset*2,
+            screenSize.height - inset*2);
+
+        setContentPane(desktopPane);
+
+
+        LogWindow logWindow = createLogWindow();
+        addWindow(logWindow);
+
+        GameWindow gameWindow = new GameWindow();
+        gameWindow.setSize(400,  400);
+        addWindow(gameWindow);
+        MenuBar menuBar = new MenuBar(this);
+        setJMenuBar(menuBar.generateMenuBar());
+        setDefaultCloseOperation(EXIT_ON_CLOSE);
+    }
+
+    protected LogWindow createLogWindow()
+    {
+        LogWindow logWindow = new LogWindow(Logger.getDefaultLogSource());
+        logWindow.setLocation(10,10);
+        logWindow.setSize(300, 800);
+        setMinimumSize(logWindow.getSize());
+        logWindow.pack();
+        Logger.debug("Протокол работает");
+        return logWindow;
+    }
+
+    protected void addWindow(JInternalFrame frame)
+    {
+        desktopPane.add(frame);
+        frame.setVisible(true);
+    }
+
+//    protected JMenuBar createMenuBar() {
+//        JMenuBar menuBar = new JMenuBar();
+//
+//        //Set up the lone menu.(Настройте одиночное меню.)
+//        JMenu menu = new JMenu("Document");
+//        menu.setMnemonic(KeyEvent.VK_D);
+//        menuBar.add(menu);
+//
+//        //Set up the first menu item.(Настройте первый пункт меню.)
+//        JMenuItem menuItem = new JMenuItem("New");
+//        menuItem.setMnemonic(KeyEvent.VK_N);
+//        menuItem.setAccelerator(KeyStroke.getKeyStroke(
+//                KeyEvent.VK_N, ActionEvent.ALT_MASK));
+//        menuItem.setActionCommand("new");
+////        menuItem.addActionListener(this);
+//        menu.add(menuItem);
+//
+//        //Set up the second menu item.(Настройте второй пункт меню.)
+//        menuItem = new JMenuItem("Quit");
+//        menuItem.setMnemonic(KeyEvent.VK_Q);
+//        menuItem.setAccelerator(KeyStroke.getKeyStroke(
+//                KeyEvent.VK_Q, ActionEvent.ALT_MASK));
+//        menuItem.setActionCommand("quit");
+////        menuItem.addActionListener(this);
+//        menu.add(menuItem);
+//
+//        return menuBar;
+//    }
+
+
+}
