+++ src/main/java/org/example/log/LogEntry.java
