+++ src/main/java/org/example/log/LogWindowSource.java
@@ -1,32 +1,35 @@
 package log;
 
+import java.util.AbstractQueue;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Queue;
+import java.util.concurrent.ArrayBlockingQueue;
 
 /**
  * Что починить:
  * 1. Этот класс порождает утечку ресурсов (связанные слушатели оказываются
  * удерживаемыми в памяти)
- * 2. Этот класс хранит активные сообщения лога, но в такой реализации он 
- * их лишь накапливает. Надо же, чтобы количество сообщений в логе было ограничено 
- * величиной m_iQueueLength (т.е. реально нужна очередь сообщений 
- * ограниченного размера) 
+ * 2. Этот класс хранит активные сообщения лога, но в такой реализации он
+ * их лишь накапливает. Надо же, чтобы количество сообщений в логе было ограничено
+ * величиной m_iQueueLength (т.е. реально нужна очередь сообщений
+ * ограниченного размера)
  */
 public class LogWindowSource
 {
     private int m_iQueueLength;
-    
-    private ArrayList<LogEntry> m_messages;
+
+    private Queue<LogEntry> m_messages;
     private final ArrayList<LogChangeListener> m_listeners;
     private volatile LogChangeListener[] m_activeListeners;
-    
-    public LogWindowSource(int iQueueLength) 
+
+    public LogWindowSource(int iQueueLength)
     {
         m_iQueueLength = iQueueLength;
-        m_messages = new ArrayList<LogEntry>(iQueueLength);
+        m_messages = new ArrayBlockingQueue<>(iQueueLength); // блокирующая очередь
         m_listeners = new ArrayList<LogChangeListener>();
     }
-    
+
     public void registerListener(LogChangeListener listener)
     {
         synchronized(m_listeners)
@@ -35,7 +38,7 @@ public class LogWindowSource
             m_activeListeners = null;
         }
     }
-    
+
     public void unregisterListener(LogChangeListener listener)
     {
         synchronized(m_listeners)
@@ -44,11 +47,16 @@ public class LogWindowSource
             m_activeListeners = null;
         }
     }
-    
+
     public void append(LogLevel logLevel, String strMessage)
     {
         LogEntry entry = new LogEntry(logLevel, strMessage);
-        m_messages.add(entry);
+        synchronized (m_messages){
+            if (m_messages.size() > m_iQueueLength){
+                m_messages.poll(); // peek() - извлекает, но не удаляет; pool() - извлекает и удаляет заголовок
+            }
+            m_messages.add(entry);
+        }
         LogChangeListener [] activeListeners = m_activeListeners;
         if (activeListeners == null)
         {
@@ -61,12 +69,13 @@ public class LogWindowSource
                 }
             }
         }
+        assert m_activeListeners != null;
         for (LogChangeListener listener : activeListeners)
         {
             listener.onLogChanged();
         }
     }
-    
+
     public int size()
     {
         return m_messages.size();
@@ -79,7 +88,7 @@ public class LogWindowSource
             return Collections.emptyList();
         }
         int indexTo = Math.min(startFrom + count, m_messages.size());
-        return m_messages.subList(startFrom, indexTo);
+        return m_messages.stream().toList().subList(startFrom, indexTo);
     }
 
     public Iterable<LogEntry> all()
